requirements.txt
# python 3.9.5
faiss-cpu==1.7.1post2
jpl.pipedreams==1.0.3
tqdm==4.62.3
scikit-learn==0.24.2 # gives memory segmentation with faiss error when version 1.0
transformers==4.10.0
tensorflow==2.6.0
numpy==1.19.5
tqdm==4.62.3
elasticsearch==7.15.1
pandas==1.3.4
pyarrow==5.0.0
datasets==1.14.0
